backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0
pydantic[email]==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
celery==5.3.4
redis==5.0.1
httpx==0.25.2
python-dotenv==1.0.0
authlib==1.2.1
itsdangerous==2.1.2

# ML libraries (for future use)
scikit-learn==1.3.2
numpy==1.25.2
pandas==2.1.3
matplotlib==3.8.2
seaborn==0.13.0